+++ src/classes/utils/Client/constructors/keepAlive.js
@@ -25,7 +25,7 @@ module.exports = {
             })
         }, sendKeepAliveInterval)
 
-        this.p.client.mpOn('keep_alive', ({ keepAliveId }) => {
+        this.p.mpOn('keep_alive', ({ keepAliveId }) => {
             keepAlivePromises[keepAliveId[1]].resolved = true;
             keepAlivePromises[keepAliveId[1]].res();
         })