+++ user/server-chunk.js
@@ -1,10 +1,15 @@
-const { Server, Chunk } = require('../');
+const { Server, Chunk, ProxyClient } = require('../');
 
 let chunk = new Chunk();
 for (let x = 0; x < 16; x++)
     for (let z = 0; z < 16; z++)
         for (let y = 0; y < 100; y++)
             chunk.setBlock('dirt', { x, y, z })
+
+let proxyClient = new ProxyClient();
+proxyClient.onPacket((name, packet) => {
+    console.log('<-', name, packet);
+})
 
 const server = new Server({
     defaultClientProperties: () => ({
@@ -23,3 +28,5 @@ server.on('connect', client => {
         for (let z = -5; z < 5; z++)
             client.chunk(chunk, { x, z });
 })
+
+server.joinProxyClient(proxyClient);